+++ examples/when_ready.conf.py
@@ -27,8 +27,8 @@ class MemoryWatch(threading.Thread):
         while True:
             for (pid, worker) in list(self.server.WORKERS.items()):
                 if self.memory_usage(pid) > self.max_mem:
-                    self.server.log.info("Pid %s killed (memory usage > %s)" % (
-                        pid, self.max_mem))
+                    self.server.log.info("Pid %s killed (memory usage > %s)", 
+                        pid, self.max_mem)
                     self.server.kill_worker(pid, signal.SIGQUIT)
             time.sleep(self.timeout)