+++ gunicorn/workers/ggevent.py
@@ -18,7 +18,7 @@ except ImportError:
     raise RuntimeError("You need gevent installed to use this worker.")
 from gevent.pool import Pool
 from gevent.server import StreamServer
-from gevent import pywsgi, wsgi
+from gevent import pywsgi
 
 import gunicorn
 from gunicorn.workers.async import AsyncWorker
@@ -51,9 +51,6 @@ class GGeventServer(StreamServer):
 
 class GeventWorker(AsyncWorker):
 
-    def __init__(self, *args, **kwargs):
-        super(GeventWorker, self).__init__(*args, **kwargs)
-
     @classmethod  
     def setup(cls):
         from gevent import monkey
@@ -76,7 +73,7 @@ class GeventWorker(AsyncWorker):
             while self.alive:
                 self.notify()
                 if self.ppid != os.getppid():
-                    self.log.info("Parent changed, shutting down: %s" % self)
+                    self.log.info("Parent changed, shutting down: %s", self)
                     break
         
                 gevent.sleep(1.0)
@@ -97,12 +94,15 @@ class GeventWorker(AsyncWorker):
         except gevent.GreenletExit:
             pass
 
-    def init_process(self):
-        #gevent doesn't reinitialize dns for us after forking
-        #here's the workaround
-        gevent.core.dns_shutdown(fail_requests=1)
-        gevent.core.dns_init()
-        super(GeventWorker, self).init_process()
+    if hasattr(gevent.core, 'dns_shutdown'):
+
+        def init_process(self):
+            #gevent 0.13 and older doesn't reinitialize dns for us after forking
+            #here's the workaround
+            gevent.core.dns_shutdown(fail_requests=1)
+            gevent.core.dns_init()
+            super(GeventWorker, self).init_process()
+
 
 class GeventBaseWorker(Worker):
     """\
@@ -139,7 +139,7 @@ class GeventBaseWorker(Worker):
                 self.notify()
             
                 if self.ppid != os.getppid():
-                    self.log.info("Parent changed, shutting down: %s" % self)
+                    self.log.info("Parent changed, shutting down: %s", self)
                     break
                 
                 gevent.sleep(1.0) 
@@ -154,32 +154,14 @@ class GeventBaseWorker(Worker):
         except:
             pass
 
-    def init_process(self):
-        #gevent doesn't reinitialize dns for us after forking
-        #here's the workaround
-        gevent.core.dns_shutdown(fail_requests=1)
-        gevent.core.dns_init()
-        super(GeventBaseWorker, self).init_process()
+    if hasattr(gevent.core, 'dns_shutdown'):
 
-        
-
-class WSGIHandler(wsgi.WSGIHandler):
-    def log_request(self, *args):
-        pass
-
-    def prepare_env(self):
-        env = super(WSGIHandler, self).prepare_env()
-        env['RAW_URI'] = self.request.uri
-        return env
-        
-        
-class WSGIServer(wsgi.WSGIServer):
-    base_env = BASE_WSGI_ENV        
-    
-class GeventWSGIWorker(GeventBaseWorker):
-    "The libevent HTTP based workers"
-    server_class = WSGIServer
-    wsgi_handler = WSGIHandler
+        def init_process(self):
+            #gevent 0.13 and older doesn't reinitialize dns for us after forking
+            #here's the workaround
+            gevent.core.dns_shutdown(fail_requests=1)
+            gevent.core.dns_init()
+            super(GeventBaseWorker, self).init_process()
 
 
 class PyWSGIHandler(pywsgi.WSGIHandler):